# server/API/api_docs.yaml
swagger: "2.0"
info:
  description: "Fast Food Fast is a food delivery service app for restaurants. It was built to bring the food store close as close as a click to where ever we may be. the product was built from scratch `using Html` , `Css` , `JavaScript` and `Node.js`"
  version: "1.0.0"
  title: "Fast Food App Documentation"
  termsOfService: "https://shegsteham.github.io/Fast-Food-Fast/UI/LICENSE.md"
  contact:
    email: "oluwasegunadepoju@gmail.com"
  license:
    name: "Licence: MIT"
    url: "https://shegsteham.github.io/Fast-Food-Fast/UI/LICENSE.md"
host: "fast-food-fast1.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /auth/login:
    post:
      tags:
      - "User"
      summary: "Login a user"
      description: ""
      operationId: "loginUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Details of user loging in"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Login Successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
        404:
          description: 'User Not Found in the Database'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
      security:
      - api_key: []
...
  /users:
    get:
      tags:
      - "User"
      summary: "Find all available User Data"
      description: ""
      operationId: "getAllUsers"
      produces:
      - "application/json"
      responses:
        200:
          description: 'Users Data retrieved successfully'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
      security:
      - api_key: []

          
    post:
      tags:
      - "User"
      summary: "Add a new User to the database"
      description: ""
      operationId: "addUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Details of new User to be added"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: 'User Sent Successfully'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
      security:
      - api_key: []
  /users/{userId}:
    get:
      tags:
      - "User"
      summary: "Find a particular User"
      description: ""
      operationId: "getParticularUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of the User to be retrieved"
        required: true
        type: "integer"
      responses:
        200:
          description: 'User retrieved successfully'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
        404:
          description: 'User Found in the Database'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
      security:
      - api_key: []

    put:
      tags:
      - "User"
      summary: "Update an existing User Data in the database"
      description: ""
      operationId: "updateUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of the User Data to be Edited"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "User object that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "Update successful"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
        410:
          description: "Requested resourse is no longer available"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
      security:
      - api_key: []

    delete:
      tags:
      - "User"
      summary: "Delete an existing user data in the database"
      description: ""
      operationId: "deleteUser"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of the user data to be deleted"
        required: true
        type: "integer"
      responses:
        200:
          description: 'User deleted successfuly'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponse"
        400:
          description: "Bad Request"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
        404:
          description: 'User Not Found in the Database'
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ApiResponseErr"
      security:
      - api_key: []
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
definitions:
...
  User:
    type: "object"
    properties:
      user_name:
        type: "string"
        example: "SwaggerUser"
      user_email:
        type: "string"
        example: "email@example.com"
      user_password:
        type: "string"
        example: 'swagger123swagger'
      user_role:
        type: "string"
        description: "user_role"
        enum:
        - "User"
        - "Admin"
    xml:
      name: "User"
...
  ApiResponse:
    type: "object"
    properties:
      status:
        type: "string"
      message:
        type: "string"
      data:
        type: "object"
  ApiResponseErr:
    type: "object"
    properties:
      success:
        type: "boolean"
        example: "false"
      message:
        type: "string"
...